Refactor _update_intersphinx_data
