+++ readthedocs/builds/tasks.py
@@ -20,13 +20,14 @@ from django.db.models import Q
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
+from sphinx.ext import intersphinx
 
 
 from readthedocs.core.resolver import resolve_path
 from readthedocs.core.utils import send_email
 from readthedocs.doc_builder.constants import DOCKER_LIMITS
+from readthedocs.sphinx_domains.models import SphinxDomain
 from readthedocs.projects.constants import LOG_TEMPLATE
-from readthedocs.projects.tasks import _update_intersphinx_data
 from readthedocs.projects.models import HTMLFile, ImportedFile, Project
 from readthedocs.projects.signals import (
     bulk_post_create,
@@ -318,3 +319,63 @@ def finish_inactive_builds():
         'Builds marked as "Terminated due inactivity": %s',
         builds_finished,
     )
+
+
+def _update_intersphinx_data(version, path, commit):
+    """
+    Update intersphinx data for this version
+
+    :param version: Version instance
+    :param path: Path to search
+    :param commit: Commit that updated path
+    """
+    object_file = os.path.join(path, 'objects.inv')
+    if not os.path.exists(object_file):
+        log.debug('No objects.inv, skipping intersphinx indexing.')
+        return
+
+    # These classes are copied from Sphinx
+    # https://git.io/fhFbI
+    class MockConfig:
+        intersphinx_timeout = None
+        tls_verify = False
+
+    class MockApp:
+        srcdir = ''
+        config = MockConfig()
+
+        def warn(self, msg):
+            log.warning('Sphinx MockApp: %s', msg)
+
+    invdata = intersphinx.fetch_inventory(MockApp(), '', object_file)
+    for key, value in sorted(invdata.items() or {}):
+        domain, _type = key.split(':')
+        for name, einfo in sorted(value.items()):
+            # project, version, url, display_name
+            # ('Sphinx', '1.7.9', 'faq.html#epub-faq', 'Epub info')
+            url = einfo[2]
+            if '#' in url:
+                doc_name, anchor = url.split(
+                    '#',
+                    # The anchor can contain ``#`` characters
+                    maxsplit=1
+                )
+            else:
+                doc_name, anchor = url, ''
+            display_name = einfo[3]
+            obj, _ = SphinxDomain.objects.get_or_create(
+                project=version.project,
+                version=version,
+                domain=domain,
+                name=name,
+                display_name=display_name,
+                type=_type,
+                doc_name=doc_name,
+                anchor=anchor,
+            )
+            if obj.commit != commit:
+                obj.commit = commit
+                obj.save()
+    SphinxDomain.objects.filter(project=version.project,
+                                version=version
+                                ).exclude(commit=commit).delete()
